Improve correctness, and efficiency
--debug now expects a debug level

Add --group for grouping entries with the same date/time

Track how many AWS API calls are used

Now defaulting to querying the log-events of log-stream starting from
the head, instead of from the tail, as that appears to require the
same or fewer API calls, most of the time. You can

Add --grep <regex> to either match log group names, or match event
messages. The pattern is expected to match the entire item.